requirements.txt
pytest
flake8
attrs>=18.1.0
mypy>=0.600
wheel>=0.31.1
tornado>=5.0.2
coveralls
coverage
pytest-cov
asyncpg>=0.15.0
testing.postgresql>=1.3.0
pytest-asyncio>=0.8.0
pytest-tornado
